+++ components/spell.lua
@@ -2,7 +2,7 @@ local AddonName, SAO = ...
 
 -- Optimize frequent calls
 local GetSpellInfo = GetSpellInfo
-local IsPlayerSpell = IsPlayerSpell
+local IsSpellKnownOrOverridesKnown = IsSpellKnownOrOverridesKnown
 
 -- List of spell IDs sharing the same name
 -- key = spell name, value = list of spell IDs
@@ -102,13 +102,8 @@ function SAO.IsSpellIdentical(self, spellID, spellName, referenceID)
 end
 
 -- Test if the player is capable of casting a specific spell
--- For most game projects, it checks the IsPlayerSpell function
--- For Season of Discovery, it adds a specific check related to runes
 function SAO.IsSpellLearned(self, spellID)
-    if IsPlayerSpell(spellID) then
-        return true;
-    end
-    if spellID >= 400000 and self.IsSoD() and self:IsRuneSpellLearned(spellID) then
+    if IsSpellKnownOrOverridesKnown(spellID) then
         return true;
     end
     return false;
@@ -151,8 +146,8 @@ function SAO.IsFakeSpell(self, spellID)
         return true
     end
 
-    if (self.IsEra() or self.IsTBC() or self.IsWrath()) and spellID == 48107 then
-        -- Mage's Heating Up does not exist in Era/TBC/Wrath
+    if (self.IsEra() or self.IsTBC() or self.IsWrath() or self.IsCata()) and spellID == 48107 then
+        -- Mage's Heating Up does not exist in Era/TBC/Wrath/Cata
         return true
     end