+++ changelog.md
@@ -1,15 +1,52 @@
 ## SpellActivationOverlay Changelog
 
-#### v1.2.1 (2024-04-xx)
+#### v1.3.1 (2024-04-xx)
+
+Shout-out to fellow developer Amanthuul. Thanks!
+
+Options:
+- New option to play a sound effect when triggering a spell alert
+- The option is enabled by default for Cataclysm, which had this sound
+- The option is disabled by default for Classic Era and Wrath
+- It would be confusing to add sounds overnight to Era and Wrath players
+- To enable or disable it, please open the options panel by entering /sao
+
+New runes for Season of Discovery:
+- New SAO: Shaman's Rolling Thunder with 7-9 Lightning Shield stacks
+- New SAO: Shaman's Tidal Waves
+- New GAB: Shaman's Earth Shock, with 7-9 Lightning Shield stacks
+- New GAB: Shaman's Healing Wave, during Tidal Waves
+- New GAB: Shaman's Lesser Healing Wave, during Tidal Waves
+
+Introducing Cataclysm flavor!
+- the addon is in very early stage, currently in alpha
+- support for Mage's Arcane Missiles
+- support for Mage's Arcane Potency
+- support for Mage's Brain Freeze
+- support for Mage's Clearcasting
+- support for Mage's Deep Freeze
+- support for Mage's Fingers of Frost
+- support for Mage's Frozen debuff
+- support for Mage's Heating Up and Hot Streak
+- support for Mage's Impact
+
+#### v1.3.0 (2024-04-17)
 
 Bump in TOC file for Season of Discovery update (Classic Era)
 
-New runes for Season of Discovery:
-- New SAO: Priest's Surge of Light
-- New GAB: Priest's Flash Heal button glows during Surge of Light
-- New GAB: Priest's Smite button glows during Surge of Light
+- New GAB: Paladin's Exorcism
+- New SAO: Priest's Surge of Light (Season of Discovery)
+- New GAB: Priest's Flash Heal button glows during Surge of Light (SoD)
+- New GAB: Priest's Smite button glows during Surge of Light (SoD)
+
+Combat-only counters:
+- Hunter's Flanking Strike now fades out after leaving combat
+- Mage's Heating Up now fades out after leaving combat
+- Paladin's Exorcism fades out after leaving combat
+- Shaman's Molten Blast now fades out after leaving combat
 
 Bug Fixes
+- Counters are no longer flagged as unavailable during GCD (Classic Era)
 - Druid's Eclipse now refreshes its visual timer when gaining a stack
 
 #### v1.2.0 (2024-03-19)